+++ examples/debug/inspyred_debug.py
@@ -3,7 +3,7 @@ import logging
 import operator
 
 from src.optim.inspyred_optimizer import DebugOptimizer
-from src.optim.evolutor import InspyredEvolution
+from src.optim.evolution import InspyredEvolution
 
 logger = logging.getLogger()
